requirements.txt
et-xmlfile==1.1.0
lxml==5.2.2
openpyxl==3.1.5
pillow==10.4.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-pptx==0.6.23
XlsxWriter==3.2.0
